requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary     
sqlalchemy
alembic             
qdrant-client
langchain
openai
pydantic
python-dotenv       
celery
redis
google-generativeai    
pymongo                 